COMMUNICATION-2657 - Inject HsmEnergyServiceImpl into protocol services Refactor the way hsm service is injected into mdc.upl Services +other small changes

+++ comu/com.energyict.mdc.device.config.rest/src/main/java/com/energyict/mdc/device/configuration/rest/impl/SecurityAccessorTypeInfo.java
@@ -5,12 +5,10 @@
 package com.energyict.mdc.device.configuration.rest.impl;
 
 import com.elster.jupiter.hsm.model.keys.SessionKeyCapability;
-import com.elster.jupiter.pki.impl.ProtocolKeyTypes;
 import com.elster.jupiter.rest.util.IdWithNameInfo;
 import com.elster.jupiter.time.rest.TimeDurationInfo;
 import com.energyict.mdc.device.configuration.rest.ExecutionLevelInfo;
 import com.energyict.mdc.device.configuration.rest.SecurityAccessorInfo;
-
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 import java.util.ArrayList;
@@ -35,6 +33,6 @@ public class SecurityAccessorTypeInfo {
     public String label;
     public SessionKeyCapability importCapability;
     public SessionKeyCapability renewCapability;
-    public short keySize;
+    public int keySize;
 
 }
